Fix integer overflow in tensor index calculation (Offset) and validate filter dimensions and shape bounds in TFLite pooling kernels.
PiperOrigin-RevId: 964564472

+++ tensorflow/compiler/mlir/lite/kernels/internal/runtime_shape.h
@@ -372,6 +372,7 @@ inline RuntimeShape DimsToShape(const mlir::Dims<4>& dims) {
 // calling ops to ensure that they perform verification checks on tensor shapes
 // if they don't support a particular behavior.
 
+// LINT.IfChange
 inline int Offset(const RuntimeShape& shape, int i0, int i1, int i2, int i3) {
   TFLITE_DCHECK_EQ(shape.DimensionsCount(), 4);
   const int* dims_data = reinterpret_cast<const int*>(shape.DimsDataUpTo5D());
@@ -383,7 +384,10 @@ inline int Offset(const RuntimeShape& shape, int i0, int i1, int i2, int i3) {
                 (i2 >= 0 && i2 < dims_data[2]));
   TFLITE_DCHECK((dims_data[3] == 0 && i3 == 0) ||
                 (i3 >= 0 && i3 < dims_data[3]));
-  return ((i0 * dims_data[1] + i1) * dims_data[2] + i2) * dims_data[3] + i3;
+  return static_cast<int>(
+      ((static_cast<int64_t>(i0) * dims_data[1] + i1) * dims_data[2] + i2) *
+          dims_data[3] +
+      i3);
 }
 
 inline int Offset(const RuntimeShape& shape, int i0, int i1, int i2, int i3,
@@ -400,9 +404,12 @@ inline int Offset(const RuntimeShape& shape, int i0, int i1, int i2, int i3,
                 (i3 >= 0 && i3 < dims_data[3]));
   TFLITE_DCHECK((dims_data[4] == 0 && i4 == 0) ||
                 (i4 >= 0 && i4 < dims_data[4]));
-  return (((i0 * dims_data[1] + i1) * dims_data[2] + i2) * dims_data[3] + i3) *
-             dims_data[4] +
-         i4;
+  return static_cast<int>(
+      (((static_cast<int64_t>(i0) * dims_data[1] + i1) * dims_data[2] + i2) *
+            dims_data[3] +
+        i3) *
+          dims_data[4] +
+      i4);
 }
 
 inline int Offset(const RuntimeShape& shape, int* index) {
